requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.9.2
python-dotenv==1.0.1
python-multipart==0.0.12

# OCR stack
opencv-python-headless==4.6.0.66  # keep this version; works with PaddleOCR 2.7
Pillow==10.4.0
numpy==1.26.4

# PDF rasterization (no PyMuPDF)
pypdfium2==4.30.0

# PaddleOCR runtime deps (subset)
shapely==2.1.2
scikit-image==0.25.2
imgaug==0.4.0
pyclipper==1.3.0.post6
lmdb==1.7.5
tqdm==4.67.1
rapidfuzz==3.14.3
lxml==6.0.2
premailer==3.10.0
openpyxl==3.1.5
attrdict==2.0.1
visualdl==2.5.3
scikit-learn==1.5.2

# Handwriting detection & TrOCR
torch==2.1.2
torchvision==0.16.2
transformers==4.36.2

# utils/tests
imagehash==4.3.1
pytest==8.3.3

